refactor: move round cooldown to roundManager
