+++ src/helpers/classicBattle/timerService.js
@@ -11,54 +11,10 @@ import { isEnabled } from "../featureFlags.js";
 import { realScheduler } from "../scheduler.js";
 import { dispatchBattleEvent } from "./orchestrator.js";
 import { createRoundTimer } from "../timers/createRoundTimer.js";
-import { computeNextRoundCooldown } from "../timers/computeNextRoundCooldown.js";
-import { attachCooldownRenderer } from "../CooldownRenderer.js";
 import { getStateSnapshot } from "./battleDebug.js";
 const IS_VITEST = typeof process !== "undefined" && !!process.env?.VITEST;
-import { exposeDebugState, readDebugState } from "./debugHooks.js";
-
-/**
- * Store controls for the pending next round. Updated by `scheduleNextRound`
- * and consumed by `onNextButtonClick` when invoked via the Next button.
- * @type {{timer: ReturnType<typeof createRoundTimer>|null, resolveReady: (()=>void)|null, ready: Promise<void>|null}|null}
- */
-let currentNextRound = null;
-
-/**
- * Schedule a fallback timeout and return its id.
- *
- * @pseudocode
- * 1. Attempt to call `setTimeout(cb, ms)`.
- * 2. Return the timer id or `null` on failure.
- *
- * @param {number} ms
- * @param {Function} cb
- * @returns {ReturnType<typeof setTimeout>|null}
- */
-export function setupFallbackTimer(ms, cb) {
-  try {
-    return setTimeout(cb, ms);
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Mark the Next button as ready and enable it.
- *
- * @pseudocode
- * 1. If `btn` exists:
- * 2.   - Set `data-next-ready` to "true".
- * 3.   - Enable the button.
- *
- * @param {HTMLButtonElement | null} btn - Next button element.
- */
-export function markNextReady(btn) {
-  if (btn) {
-    btn.dataset.nextReady = "true";
-    btn.disabled = false;
-  }
-}
+import { getNextRoundControls } from "./roundManager.js";
+export { getNextRoundControls } from "./roundManager.js";
 
 // Skip handler utilities moved to skipHandler.js
 
@@ -172,8 +128,8 @@ export async function cancelTimerOrAdvance(_btn, timer, resolveReady) {
  * Handle clicks on the Next button by delegating to helper functions.
  *
  * @param {MouseEvent} _evt - Click event.
- * @param {{timer: {stop: () => void} | null, resolveReady: (() => void) | null}} [controls=currentNextRound]
- * - Timer controls returned from `scheduleNextRound`.
+ * @param {{timer: {stop: () => void} | null, resolveReady: (() => void) | null}} [controls=getNextRoundControls()]
+ * - Timer controls returned from `startCooldown`.
  * @example
  * const controls = { timer, resolveReady };
  * await onNextButtonClick(new MouseEvent("click"), controls);
@@ -198,7 +154,8 @@ export async function cancelTimerOrAdvance(_btn, timer, resolveReady) {
  * @pseudocode
  * 1. TODO: Add pseudocode
  */
-export async function onNextButtonClick(_evt, { timer, resolveReady } = currentNextRound ?? {}) {
+export async function onNextButtonClick(_evt, controls = getNextRoundControls()) {
+  const { timer = null, resolveReady = null } = controls || {};
   const btn = document.getElementById("next-button");
   if (!btn) return;
   if (btn.dataset.nextReady === "true") {
@@ -208,34 +165,7 @@ export async function onNextButtonClick(_evt, { timer, resolveReady } = currentN
   await cancelTimerOrAdvance(btn, timer, resolveReady);
 }
 
-/**
- * Expose current next-round controls for helpers like `setupNextButton`.
- *
- * @returns {{timer: ReturnType<typeof createRoundTimer>|null, resolveReady: (()=>void)|null, ready: Promise<void>|null}|null}
- */
-/**
- * @summary TODO: Add summary
- * @pseudocode
- * 1. TODO: Add pseudocode
- */
-/**
- * @summary TODO: Add summary
- * @pseudocode
- * 1. TODO: Add pseudocode
- */
-/**
- * @summary TODO: Add summary
- * @pseudocode
- * 1. TODO: Add pseudocode
- */
-/**
- * @summary TODO: Add summary
- * @pseudocode
- * 1. TODO: Add pseudocode
- */
-export function getNextRoundControls() {
-  return currentNextRound;
-}
+// `getNextRoundControls` re-exported from roundManager.js
 
 /**
  * Helper to force auto-select and dispatch outcome on timer error or drift.
@@ -438,216 +368,4 @@ export function handleStatSelectionTimeout(
       }
     }
   }, timeoutMs);
-}
-
-/**
- * Handle expiration of the next-round cooldown.
- *
- * @pseudocode
- * 1. Clear the skip handler and scoreboard timer.
- * 2. Enable the Next button and mark it as ready.
- * 3. Wait for the battle state to reach "cooldown".
- * 4. Dispatch "ready" and wait for handlers to complete.
- * 5. Mark the button as ready again, update the debug panel, and resolve the ready promise.
- *
- * @param {{resolveReady: (() => void) | null}} controls - Timer controls.
- * @param {HTMLButtonElement | null} btn - Next button element.
- * @returns {Promise<void>}
- */
-export async function handleNextRoundExpiration(controls, btn) {
-  setSkipHandler(null);
-  scoreboard.clearTimer();
-  markNextReady(btn);
-  await new Promise((resolve) => {
-    try {
-      const state = getStateSnapshot().state;
-      if (!state || state === "cooldown") {
-        resolve();
-        return;
-      }
-    } catch {}
-    const handler = (e) => {
-      try {
-        if (e.detail?.to === "cooldown") {
-          offBattleEvent("battleStateChange", handler);
-          resolve();
-        }
-      } catch {}
-    };
-    onBattleEvent("battleStateChange", handler);
-  });
-  try {
-    try {
-      const IS_VITEST = typeof process !== "undefined" && !!process.env?.VITEST;
-      if (!IS_VITEST) {
-        console.log("DEBUG: handleNextRoundExpiration dispatching 'ready'");
-      }
-    } catch {}
-    await dispatchBattleEvent("ready");
-  } catch {}
-  markNextReady(btn);
-  try {
-    const { updateDebugPanel } = await import("./uiHelpers.js");
-    updateDebugPanel();
-  } catch {}
-  if (typeof controls.resolveReady === "function") {
-    controls.resolveReady();
-  }
-}
-
-/**
- * Create an expiration handler for the next-round cooldown timer.
- *
- * @pseudocode
- * 1. Ignore subsequent calls after the first invocation.
- * 2. Delegate to `handleNextRoundExpiration(controls, btn)`.
- *
- * @param {{resolveReady: (() => void) | null}} controls - Timer controls.
- * @param {HTMLButtonElement | null} btn - Next button element.
- * @returns {() => Promise<void>} Expiration handler.
- */
-function createNextRoundOnExpired(controls, btn) {
-  let expired = false;
-  return () => {
-    if (expired) return;
-    expired = true;
-    return handleNextRoundExpiration(controls, btn);
-  };
-}
-
-/**
- * Prepare and return the cooldown timer for the next round.
- *
- * @pseudocode
- * 1. Create a round timer and attach `CooldownRenderer`.
- * 2. Register the provided expiration handler.
- * 3. Forward drift events to a fallback message.
- * 4. Register a skip handler that stops the timer.
- *
- * @param {{timer: ReturnType<typeof createRoundTimer>|null}} controls - Timer controls.
- * @param {HTMLButtonElement | null} btn - Next button element.
- * @param {number} cooldownSeconds - Cooldown duration in seconds.
- * @param {() => Promise<void>} onExpired - Expiration handler.
- * @returns {ReturnType<typeof createRoundTimer>} Prepared timer instance.
- */
-function prepareNextRoundTimer(controls, btn, cooldownSeconds, onExpired) {
-  const timer = createRoundTimer();
-  attachCooldownRenderer(timer, cooldownSeconds);
-  timer.on("expired", onExpired);
-  timer.on("drift", () => {
-    const msgEl = document.getElementById("round-message");
-    if (msgEl && msgEl.textContent) {
-      showSnackbar("Waiting…");
-    } else {
-      scoreboard.showMessage("Waiting…");
-    }
-  });
-  controls.timer = timer;
-  setSkipHandler(() => {
-    try {
-      console.warn("[test] skip: stop nextRoundTimer");
-    } catch {}
-    controls.timer?.stop();
-  });
-  return timer;
-}
-
-/**
- * Enable the Next Round button after a cooldown period.
- *
- * @pseudocode
- * 1. If the match ended, resolve immediately.
- * 2. Determine cooldown seconds via `computeNextRoundCooldown` (minimum 1s).
- * 3. Locate `#next-button` and reset the button state.
- * 4. Attach `CooldownRenderer` and wire expiration callbacks via helpers.
- * 5. Register a skip handler that logs the skip (for tests) and stops the timer.
- * 6. Start the timer and schedule a fallback with `setupFallbackTimer`.
- *
- * @param {{matchEnded: boolean}} result - Result from a completed round.
- * @returns {{
- *   ready: Promise<void>,
- *   timer: ReturnType<typeof createRoundTimer> | null,
- *   resolveReady: (() => void) | null
- * }} Controls for the scheduled next round.
- */
-export function scheduleNextRound(result, scheduler = realScheduler) {
-  logScheduleNextRound(result);
-  const controls = createNextRoundControls();
-  if (result.matchEnded) {
-    setSkipHandler(null);
-    if (controls.resolveReady) controls.resolveReady();
-    currentNextRound = controls;
-    return controls;
-  }
-  const btn = document.getElementById("next-button");
-  if (btn) {
-    btn.disabled = false;
-    delete btn.dataset.nextReady;
-  }
-  const cooldownSeconds = computeNextRoundCooldown();
-  wireNextRoundTimer(controls, btn, cooldownSeconds, scheduler);
-  currentNextRound = controls;
-  return controls;
-}
-
-function logScheduleNextRound(result) {
-  try {
-    const { state: s } = getStateSnapshot();
-    const count = (readDebugState("scheduleNextRoundCount") || 0) + 1;
-    exposeDebugState("scheduleNextRoundCount", count);
-    if (!IS_VITEST)
-      console.warn(
-        `[test] scheduleNextRound call#${count}: state=${s} matchEnded=${!!result?.matchEnded}`
-      );
-  } catch {}
-}
-
-function createNextRoundControls() {
-  const controls = { timer: null, resolveReady: null, ready: null };
-  controls.ready = new Promise((resolve) => {
-    controls.resolveReady = () => {
-      emitBattleEvent("nextRoundTimerReady");
-      resolve();
-      controls.resolveReady = null;
-    };
-  });
-  return controls;
-}
-
-function wireNextRoundTimer(controls, btn, cooldownSeconds, scheduler) {
-  // Do not skip scheduling based on current state; roundResolved may fire
-  // while the machine is still transitioning. Scheduling early is safe — the
-  // expiration handler checks the live state before dispatching 'ready'.
-  const timer = createRoundTimer();
-  attachCooldownRenderer(timer, cooldownSeconds);
-  let expired = false;
-  const onExpired = () => {
-    if (expired) return;
-    expired = true;
-    return handleNextRoundExpiration(controls, btn);
-  };
-  timer.on("expired", onExpired);
-  timer.on("drift", () => {
-    const msgEl = document.getElementById("round-message");
-    if (msgEl && msgEl.textContent) {
-      showSnackbar("Waiting…");
-    } else {
-      scoreboard.showMessage("Waiting…");
-    }
-  });
-  controls.timer = timer;
-  setSkipHandler(() => {
-    try {
-      console.warn("[test] skip: stop nextRoundTimer");
-    } catch {}
-    controls.timer?.stop();
-  });
-
-  // Start engine-backed countdown on the next tick.
-  scheduler.setTimeout(() => controls.timer.start(cooldownSeconds), 0);
-  // Fallback to ensure expiration when the engine isn't running.
-  try {
-    const ms = Math.max(0, Number(cooldownSeconds) * 1000) + 10;
-    setupFallbackTimer(ms, onExpired);
-  } catch {}
 }
