+++ tests/helpers/classicBattle/debugPanel.test.js
@@ -34,10 +34,14 @@ vi.mock("../../../src/helpers/battle/index.js", () => ({ showResult: vi.fn() }))
 vi.mock("../../../src/helpers/classicBattle/selectionHandler.js", () => ({
   handleStatSelection: vi.fn()
 }));
-vi.mock("../../../src/helpers/classicBattle/timerService.js", () => ({
-  onNextButtonClick: vi.fn(),
+vi.mock("../../../src/helpers/classicBattle/roundManager.js", () => ({
   getNextRoundControls: vi.fn(),
-  setupFallbackTimer: vi.fn((ms, cb) => setTimeout(cb, ms))
+  setupFallbackTimer: vi.fn((ms, cb) => setTimeout(cb, ms)),
+  startCooldown: vi.fn(),
+  handleReplay: vi.fn()
+}));
+vi.mock("../../../src/helpers/classicBattle/timerService.js", () => ({
+  onNextButtonClick: vi.fn()
 }));
 vi.mock("../../../src/helpers/stats.js", () => ({ loadStatNames: vi.fn() }));
 vi.mock("../../../src/helpers/viewportDebug.js", () => ({ toggleViewportSimulation: vi.fn() }));
