+++ tests/helpers/uiHelpers.resetBattleUI.test.js
@@ -1,8 +1,7 @@
 import { describe, it, expect, beforeEach, vi } from "vitest";
 
 vi.mock("../../src/helpers/classicBattle/timerService.js", () => ({
-  onNextButtonClick: vi.fn(),
-  setupFallbackTimer: vi.fn((ms, cb) => setTimeout(cb, ms))
+  onNextButtonClick: vi.fn()
 }));
 
 vi.mock("../../src/helpers/setupScoreboard.js", () => ({
